tool_sroll_feed/requirements.txt
playwright
playwright-stealth
shazamio
openai-whisper
pydub
webrtcvad
requests
librosa
sqlmodel
python-dotenv
ffmpeg-python
pandas
openpyxl
